backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Pomodoro backend"
authors = ["Carter Johnson <carterojohnson@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
